refactor: main - refactor code

+++ Patient.API/Patient.Core/Repositories/PatientRepository.cs
@@ -30,7 +30,7 @@ public class PatientRepository : IRepository<Models.Patient>
             .FirstOrDefaultAsync(p => p.Id == id);
     }
 
-    public async Task AddAsync(Core.Models.Patient? patient)
+    public async Task AddAsync(Core.Models.Patient patient)
     {
         await _context.Patients.AddAsync(patient);
         await _context.SaveChangesAsync();
@@ -45,7 +45,9 @@ public class PatientRepository : IRepository<Models.Patient>
     public async Task DeleteAsync(Guid id)
     {
         var patient = await GetByIdAsync(id);
-        _context.Patients.Remove(patient);
+        if (patient != null) 
+            _context.Patients.Remove(patient);
+        
         await _context.SaveChangesAsync();
     }
 
@@ -54,22 +56,23 @@ public class PatientRepository : IRepository<Models.Patient>
         return await _context.Patients.AnyAsync(p => p.Id == id);
     }
     
-    public async Task<List<Models.Patient?>> SearchByBirthDateAsync(DateTime? startDate, DateTime? endDate, string prefix)
+    public async Task<List<Models.Patient>> SearchByBirthDateAsync(DateTime? startDate, DateTime? endDate,
+        string prefix)
     {
         var query = _context.Patients.AsQueryable();
 
         if (prefix == DateConstants.NotEqualPrefix)
         {
             if (startDate == endDate && startDate.HasValue)
-                query = query.Where(p => p != null && p.BirthDate !=  startDate.Value.Date);
+                query = query.Where(p => p.BirthDate !=  startDate.Value.Date);
         }
         else
         {
             if (startDate.HasValue)
-                query = query.Where(p => p != null && p.BirthDate >= startDate.Value);
+                query = query.Where(p => p.BirthDate >= startDate.Value);
         
             if (endDate.HasValue)
-                query = query.Where(p => p != null && p.BirthDate <= endDate.Value);
+                query = query.Where(p => p.BirthDate <= endDate.Value);
         }
         
         return await query.ToListAsync();
